Feature: Added facebook oauth login and styling to the application

+++ views.py
@@ -14,9 +14,9 @@ import requests
 
 app = Flask(__name__)
 
-# CLIENT_ID = json.loads(
-#     open('client_secrets.json', 'r').read())['web']['client_id']
-# APPLICATION_NAME = "Item Category"
+CLIENT_ID = json.loads(
+	open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Item Catalog"
 
 
 # Connect to Database and create database session
@@ -25,6 +25,239 @@ Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
+# Create anti-forgery state token
+@app.route('/login')
+def showLogin():
+	state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+					for x in xrange(32))
+	login_session['state'] = state
+	# return "The current session state is %s" % login_session['state']
+	return render_template('login.html', STATE=state)
+
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+	if request.args.get('state') != login_session['state']:
+		response = make_response(json.dumps('Invalid state parameter.'), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+	access_token = request.data
+	print "access token received %s " % access_token
+
+
+	app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+		'web']['app_id']
+	app_secret = json.loads(
+		open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+	url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+		app_id, app_secret, access_token)
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[1]
+
+
+	# Use token to get user info from API
+	userinfo_url = "https://graph.facebook.com/v2.8/me"
+	'''
+		Due to the formatting for the result from the server token exchange we have to
+		split the token first on commas and select the first index which gives us the key : value
+		for the server access token then we split it on colons to pull out the actual token value
+		and replace the remaining quotes with nothing so that it can be used directly in the graph
+		api calls
+	'''
+	token = result.split(',')[0].split(':')[1].replace('"', '')
+
+	url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[1]
+	# print "url sent for API access:%s"% url
+	# print "API JSON result: %s" % result
+	data = json.loads(result)
+	login_session['provider'] = 'facebook'
+	login_session['username'] = data["name"]
+	login_session['email'] = data["email"]
+	login_session['facebook_id'] = data["id"]
+
+	# The token must be stored in the login_session in order to properly logout
+	login_session['access_token'] = token
+
+	# Get user picture
+	url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[1]
+	data = json.loads(result)
+
+	login_session['picture'] = data["data"]["url"]
+
+	# see if user exists
+	user_id = getUserID(login_session['email'])
+	if not user_id:
+		user_id = createUser(login_session)
+	login_session['user_id'] = user_id
+
+	output = ''
+	output += '<h1>Welcome, '
+	output += login_session['username']
+
+	output += '!</h1>'
+	output += '<img src="'
+	output += login_session['picture']
+	output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+
+	flash("Now logged in as %s" % login_session['username'])
+	return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+	facebook_id = login_session['facebook_id']
+	# The access token must me included to successfully logout
+	access_token = login_session['access_token']
+	url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+	h = httplib2.Http()
+	result = h.request(url, 'DELETE')[1]
+	return "you have been logged out"
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+	# Validate state token
+	if request.args.get('state') != login_session['state']:
+		response = make_response(json.dumps('Invalid state parameter.'), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+	# Obtain authorization code
+	code = request.data
+
+	try:
+		# Upgrade the authorization code into a credentials object
+		oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+		oauth_flow.redirect_uri = 'postmessage'
+		credentials = oauth_flow.step2_exchange(code)
+	except FlowExchangeError:
+		response = make_response(
+			json.dumps('Failed to upgrade the authorization code.'), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+
+	# Check that the access token is valid.
+	access_token = credentials.access_token
+	url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+		   % access_token)
+	h = httplib2.Http()
+	result = json.loads(h.request(url, 'GET')[1])
+	# If there was an error in the access token info, abort.
+	if result.get('error') is not None:
+		response = make_response(json.dumps(result.get('error')), 500)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+
+	# Verify that the access token is used for the intended user.
+	gplus_id = credentials.id_token['sub']
+	if result['user_id'] != gplus_id:
+		response = make_response(
+			json.dumps("Token's user ID doesn't match given user ID."), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+
+	# Verify that the access token is valid for this app.
+	if result['issued_to'] != CLIENT_ID:
+		response = make_response(
+			json.dumps("Token's client ID does not match app's."), 401)
+		print "Token's client ID does not match app's."
+		response.headers['Content-Type'] = 'application/json'
+		return response
+
+	stored_access_token = login_session.get('access_token')
+	stored_gplus_id = login_session.get('gplus_id')
+	if stored_access_token is not None and gplus_id == stored_gplus_id:
+		response = make_response(json.dumps('Current user is already connected.'),
+								 200)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+
+	# Store the access token in the session for later use.
+	login_session['access_token'] = credentials.access_token
+	login_session['gplus_id'] = gplus_id
+
+	# Get user info
+	userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+	params = {'access_token': credentials.access_token, 'alt': 'json'}
+	answer = requests.get(userinfo_url, params=params)
+
+	data = answer.json()
+
+	login_session['username'] = data['name']
+	login_session['picture'] = data['picture']
+	login_session['email'] = data['email']
+	# ADD PROVIDER TO LOGIN SESSION
+	login_session['provider'] = 'google'
+
+	# see if user exists, if it doesn't make a new one
+	user_id = getUserID(data["email"])
+	if not user_id:
+		user_id = createUser(login_session)
+	login_session['user_id'] = user_id
+
+	output = ''
+	output += '<h1>Welcome, '
+	output += login_session['username']
+	output += '!</h1>'
+	output += '<img src="'
+	output += login_session['picture']
+	output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+	flash("you are now logged in as %s" % login_session['username'])
+	print "done!"
+	return output
+
+# User Helper Functions
+
+
+def createUser(login_session):
+	newUser = User(name=login_session['username'], email=login_session[
+				   'email'], picture=login_session['picture'])
+	session.add(newUser)
+	session.commit()
+	user = session.query(User).filter_by(email=login_session['email']).one()
+	return user.id
+
+
+def getUserInfo(user_id):
+	user = session.query(User).filter_by(id=user_id).one()
+	return user
+
+
+def getUserID(email):
+	try:
+		user = session.query(User).filter_by(email=email).one()
+		return user.id
+	except:
+		return None
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+	# Only disconnect a connected user.
+	access_token = login_session.get('access_token')
+	if access_token is None:
+		response = make_response(
+			json.dumps('Current user not connected.'), 401)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+	url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+	h = httplib2.Http()
+	result = h.request(url, 'GET')[0]
+	if result['status'] == '200':
+		response = make_response(json.dumps('Successfully disconnected.'), 200)
+		response.headers['Content-Type'] = 'application/json'
+		return response
+	else:
+		response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+		response.headers['Content-Type'] = 'application/json'
+		return response
+
 
 
 #JSON APIs to view Category information
@@ -41,19 +274,27 @@ def categoryItemsJson(category_id):
 		category_id=category_id).all()
 	return jsonify(categoryItems=[i.serialize for i in category_items])    
 
+
+
+
 # route to get all categories
 @app.route('/')
 @app.route('/categories')
 def showCategories():
 	categories = session.query(Category).all()
-	return render_template('publicCategory.html', categories=categories)
+	if 'username' not in login_session:
+		return render_template('publicCategory.html', categories=categories)
+	else:
+		return render_template('categories.html', categories=categories)
 
 # create a new category
 @app.route('/category/new', methods=['GET', 'POST'])
 def newCategory():
+	if 'username' not in login_session:
+		return redirect('/login')
 	if request.method == 'POST':
-		newCategory = Category(name=request.form['name'],
-		 description=request.form['description'])
+		newCategory = Category(name=request.form['name'], 
+			user_id=login_session['user_id'])
 		session.add(newCategory)
 		flash('New Category %s Successfully Created' % newCategory.name)
 		session.commit()
@@ -66,6 +307,10 @@ def newCategory():
 def editCategory(category_id):
 	editedCategory = session.query(
 		Category).filter_by(id=category_id).one()
+	if 'username' not in login_session:
+		return redirect('/login')
+	if editedCategory.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to edit this  category. Please create your own  category in order to edit.');}</script><body onload='myFunction()'>"
 	if request.method == 'POST':
 		if request.form['name']:
 			editedCategory.name = request.form['name']
@@ -79,6 +324,10 @@ def editCategory(category_id):
 def deleteCategory(category_id):
 	categoryToDelete = session.query(
 		Category).filter_by(id=category_id).one()
+	if 'username' not in login_session:
+		return redirect('/login')
+	if  categoryToDelete.user_id != login_session['user_id']:
+		return "<script>function myFunction() {alert('You are not authorized to delete this  category. Please create your own  category in order to delete.');}</script><body onload='myFunction()'>"
 	if request.method == 'POST':
 		session.delete(categoryToDelete)
 		flash('%s Successfully Deleted' % categoryToDelete.name)
@@ -93,22 +342,31 @@ def deleteCategory(category_id):
 @app.route('/categories/<int:category_id>/category_item')
 def showCategoryItems(category_id):
 	category = session.query(Category).filter_by(id=category_id).one()
+	creator = getUserInfo(category.user_id)
 	categoryItems = session.query(CategoryItem).filter_by(
 		category_id=category_id).all()
-	return render_template(
-		'publicCategoryItem.html', category=category, category_items=categoryItems, category_id=category_id)
+	if 'username' not in login_session or creator.id != login_session['user_id']:
+		return render_template(
+			'publicCategoryItem.html', items=categoryItems, category_id=category_id, creator=creator)
+	else:
+		return render_template(
+		'categoryItems.html', items=categoryItems, category_id=category_id, category = category, creator=creator)
 
 
 # route to create a new category item
 @app.route('/category/<int:category_id>/category_item/new/', methods=['GET', 'POST'])
 def newCategoryItem(category_id):
+	if 'username' not in login_session:
+		return redirect('/login')
 	category = session.query(Category).filter_by(id=category_id).one()
-	if request.method == 'POST':
+	if login_session['user_id'] !=  category.user_id:
+		return "<script>function myFunction() {alert('You are not authorized to add menu items to this  category. Please create your own  category in order to add items.');}</script><body onload='myFunction()'>"
+		if request.method == 'POST':
 			newCategoryItem = CategoryItem(name=request.form['name'], description=request.form['description'], price=request.form[
-							   'price'], category_id=category_id)
-			session.add(newItem)
+							   'price'], category_id=category_id, user_id= category.user_id)
+			session.add(newCategoryItem)
 			session.commit()
-			flash('New Category %s Item Successfully Created' % (newItem.name))
+			flash('New Category %s Item Successfully Created' % (newCategoryItem.name))
 			return redirect(url_for('showCategoryItems', category_id=category_id))
 	else:
 		return render_template('newCategoryItem.html', category_id=category_id)
@@ -116,8 +374,12 @@ def newCategoryItem(category_id):
 # route to edit a category item
 @app.route('/category/<int:category_id>/<int:category_item_id>/edit/', methods=['GET', 'POST'])
 def editCategoryItem(category_id, category_item_id):
+	if 'username' not in login_session:
+		return redirect('/login')
 	editedItem = session.query(CategoryItem).filter_by(id=category_item_id).one()
 	category = session.query(Category).filter_by(id=category_id).one()
+	if login_session['user_id'] !=  category.user_id:
+		return "<script>function myFunction() {alert('You are not authorized to edit menu items to this  category. Please create your own  category in order to edit items.');}</script><body onload='myFunction()'>"
 	if request.method == 'POST':
 		if request.form['name']:
 			editedItem.name = request.form['name']
@@ -125,8 +387,6 @@ def editCategoryItem(category_id, category_item_id):
 			editedItem.description = request.form['description']
 		if request.form['price']:
 			editedItem.price = request.form['price']
-		if request.form['course']:
-			editedItem.course = request.form['course']
 		session.add(editedItem)
 		session.commit()
 		flash('Item Successfully Edited')
@@ -139,20 +399,45 @@ def editCategoryItem(category_id, category_item_id):
 @app.route('/categories/<int:category_id>/<int:category_item_id>/delete/',
 		   methods=['GET', 'POST'])
 def deleteCategoryItem(category_id, category_item_id):
+	if 'username' not in login_session:
+		return redirect('/login')
 	deletedItem = session.query(CategoryItem).filter_by(id=category_item_id).one()
+	if login_session['user_id'] !=  category.user_id:
+		return "<script>function myFunction() {alert('You are not authorized to delete menu items to this  category. Please create your own  category in order to delete items.');}</script><body onload='myFunction()'>"
 	if request.method == 'POST':
-		if request.form['name']:
-			deletedItem.name = request.form['name']
 		session.delete(deletedItem)
 		session.commit()
 		flash("Category Item has been deleted!")
-
-		return redirect(url_for('category', category_id=category_id))
+		return redirect(url_for('showCategories', category_id=category_id))
 	else:
 		# USE THE RENDER_TEMPLATE FUNCTION BELOW TO SEE THE VARIABLES YOU
 		# SHOULD USE IN YOUR EDITITEM TEMPLATE
 		return render_template(
 			'deletedCategoryItem.html', category_id=category_id, category_item_id=category_item_id, item=deletedItem)
+
+
+# Disconnect based on provider
+@app.route('/disconnect')
+def disconnect():
+	if 'provider' in login_session:
+		if login_session['provider'] == 'google':
+			gdisconnect()
+			del login_session['gplus_id']
+			del login_session['access_token']
+		if login_session['provider'] == 'facebook':
+			fbdisconnect()
+			del login_session['facebook_id']
+		del login_session['username']
+		del login_session['email']
+		del login_session['picture']
+		del login_session['user_id']
+		del login_session['provider']
+		flash("You have successfully been logged out.")
+		return redirect(url_for('showCategories'))
+	else:
+		flash("You were not logged in")
+		return redirect(url_for('showCategories'))
+
 
 if __name__ == '__main__':
 	app.secret_key = 'super_secret_key'
